refactor(Resizer.tsx): 컴포넌트로 분리
